fixes #23
added in various fixes to the server, cli, and db.

updated some of the errors to try and pinpoint where they were
originating.

modified the encrypted wrapper calls to use the new load_create method.

modified the db init to use bigint for auth_password since the version
types use u64. may change this in the future since it is some what of a
pain the ass to deal with the constant conversions.

added data directory to .gitignore

fixed issue with route path not being parsed correctly

+++ rfs-db/setup/postgres/init.sql
@@ -9,7 +9,7 @@ create table users (
 
 create table auth_password (
     user_id bigint not null primary key references users(id),
-    version int not null default 0,
+    version bigint not null default 0,
     hash varchar not null
 );
